fixed some layout in the dashboard and added download pdf function in Due Medication and Patient List
