+++ app/api/patients-api/patientList.api.ts
@@ -10,13 +10,16 @@ export async function searchPatientList(
   currentPage: number,
   sortBy?: string,
   sortOrder?: "ASC" | "DESC",
+  perPage?: number,
   router?: any // Pass router instance as a parameter
+
 ): Promise<any> {
   const requestData = {
     term: term,
     page: currentPage,
     sortBy: sortBy,
     sortOrder: sortOrder,
+    perPage: perPage,
   };
   try {
     const accessToken = getAccessToken();
