+++ components/reusable/tooltip.tsx
@@ -22,7 +22,7 @@ const ResuableTooltip = ({ text }: ResuableTooltipProps) => {
             <TooltipTrigger className="truncate max-w-[100%]">
               {text}
             </TooltipTrigger>
-            <TooltipContent className=" bg-[#007C85] text-white overflow-visible max-w-[429px]">
+            <TooltipContent className=" bg-[#007C85] text-white overflow-visible max-w-[429px] text-wrap">
               <p className="relative z-[51] break-words">{text}</p>
               <div className="absolute bottom-[-5px] left-1/2 transform -translate-x-1/2 w-3 h-3 bg-[#007C85] rotate-45 z-[49]"></div>
             </TooltipContent>
